Split DurationInput.test.ts into focused modules for better maintainability
**Strategy: Logical test splitting with shared utilities**

**Files created:**
- DurationInput.basic.test.ts (460 lines) - Basic functionality, rendering, navigation, configuration
- DurationInput.validation.test.ts (297 lines) - Input validation, error handling, edge cases
- duration-input-test-helpers.ts (239 lines) - Shared test utilities, mock data, helper functions

**Original file:** DurationInput.test.ts (1179 lines) → **Total: 996 lines** (-183 lines, -15%)

**Benefits:**
- Improved test organization by functional area
- Shared utilities eliminate code duplication
- Better test maintainability and readability
- All 86 original test functions preserved with identical functionality
- Easier to locate and modify specific test scenarios

**Testing:** All 86 DurationInput tests pass successfully
